Update estimation of occurences GWexPrec_lib.r
When filtering days that are not in the concerned period, one should also make sure that the days "t-1",..."t-nlag+1" are in the period. This can be done by replacing with NA days that are not in the period in the vector of precipitation vec.prec. 

There is no need to filter the first nlag days.